feat(web): modularize asset and contract detail sections
