+++ apps/explorer-web/src/app/[locale]/[network]/(explorer)/asset/[slug]/page.tsx
@@ -1,9 +1,10 @@
 import { Metadata } from "next";
 import { AssetContent } from "./asset-content";
 import { parseAssetSlug } from "@/lib/utils";
+import { JsonLd } from "@/components/common/json-ld";
 
 type Props = {
-  params: Promise<{ slug: string }>;
+  params: Promise<{ slug: string; network: string }>;
 };
 
 export async function generateMetadata({ params }: Props): Promise<Metadata> {
@@ -43,6 +44,39 @@ export async function generateMetadata({ params }: Props): Promise<Metadata> {
 }
 
 export default async function AssetPage({ params }: Props) {
-  const { slug } = await params;
-  return <AssetContent slug={slug} />;
+  const { slug, network } = await params;
+  const parsed = parseAssetSlug(slug);
+
+  const isNative = parsed?.issuer === "native";
+  const code = parsed?.code ?? slug;
+  const issuer = parsed?.issuer ?? "";
+
+  const jsonLd = isNative
+    ? {
+        "@context": "https://schema.org",
+        "@type": "Currency",
+        name: "Stellar Lumens",
+        alternateName: "XLM",
+        description: "XLM is the native asset of the Stellar blockchain network.",
+        url: "https://stellar.org",
+      }
+    : {
+        "@context": "https://schema.org",
+        "@type": "FinancialProduct",
+        name: `${code} — Stellar Asset`,
+        description: `${code} is a tokenized asset issued on the Stellar ${network} network by ${issuer.slice(0, 6)}…${issuer.slice(-6)}.`,
+        provider: {
+          "@type": "Organization",
+          identifier: issuer,
+          name: `Issuer ${issuer.slice(0, 6)}…${issuer.slice(-6)}`,
+        },
+        identifier: slug,
+      };
+
+  return (
+    <>
+      <JsonLd data={jsonLd} />
+      <AssetContent slug={slug} />
+    </>
+  );
 }
