+++ apps/explorer-web/src/app/[locale]/[network]/(explorer)/contract/[id]/page.tsx
@@ -1,8 +1,9 @@
 import { Metadata } from "next";
 import { ContractContent } from "./contract-content";
+import { JsonLd } from "@/components/common/json-ld";
 
 type Props = {
-  params: Promise<{ id: string }>;
+  params: Promise<{ id: string; network: string }>;
 };
 
 export async function generateMetadata({ params }: Props): Promise<Metadata> {
@@ -26,6 +27,27 @@ export async function generateMetadata({ params }: Props): Promise<Metadata> {
 }
 
 export default async function ContractPage({ params }: Props) {
-  const { id } = await params;
-  return <ContractContent id={id} />;
+  const { id, network } = await params;
+
+  const jsonLd = {
+    "@context": "https://schema.org",
+    "@type": "SoftwareApplication",
+    identifier: id,
+    name: `Soroban Contract ${id.slice(0, 6)}…${id.slice(-6)}`,
+    description: `A Soroban smart contract deployed on the Stellar ${network} network. Explore events, storage, and code.`,
+    applicationCategory: "SmartContract",
+    operatingSystem: "Stellar Network",
+    author: {
+      "@type": "Organization",
+      name: "Stellar Network",
+      url: "https://stellar.org",
+    },
+  };
+
+  return (
+    <>
+      <JsonLd data={jsonLd} />
+      <ContractContent id={id} />
+    </>
+  );
 }
