+++ apps/explorer-web/src/components/contracts/index.ts
@@ -1,3 +1,4 @@
 export * from "./contract-verification";
 export * from "./contract-event-details";
 export * from "./contract-transactions";
+export * from "./token-gallery";
